Начало физики

+++ Asteroid.h
@@ -2,7 +2,7 @@
 #include "GameObject.h"
 class Asteroid : public GameObject{
 public:
-	Asteroid(double speed[2], int position[2], int size[2], int type, int id) : GameObject(speed, position, size, type, id) {
+	Asteroid(double speed[2], float position[2], int size[2], int type, int id) : GameObject(speed, position, size, type, id) {
 		return;
 	}
